// src/app/models/peticion.model.ts
import { ClienteVM } from './cliente.model';

// Generated by https://quicktype.io

export interface Peticion {
  idSolicitudServicio: number;
  idPersona:           number;
  idProveedor:         number;
  servicio:            string;
  servicioDesc: string;
  cliente:             null;
  latitud:             number;
  longitud:            number;
  accuracy:            number;
  patente:             null;
  fechaSolicitud:      string;
  fechaSolicitudUtc:   string;
  fechaCreacion:       string;
  fechaCreacionUtc:    string;
  idEmeMbdb:           number;
  idSolMbdb:           number;
  emePerArchivo:       any[];
  emePerPosicion:      any[];
  persona:             ClienteVM;
  proveedor:           null;
  idProveedorWeb:      number;
  idUsuarioWeb:        number;
  proveedorWeb:        null;
}